requirements.txt
fastapi>=0.111.0,<0.112.0
twilio>=9.7.2,<10.0.0
vocode>=0.1.113,<0.2.0
langchain-groq
python-dotenv>=1.1.1,<2.0.0
httpx>=0.27.0,<0.28.0
numpy>=1.26.4,<2.0.0
uvicorn>=0.30.0,<0.31.0
deepgram-sdk>=3.5.0,<4.0.0
pydantic>=2.7.2,<3.0.0
langchain
# langchain-core
langchain-community
vocode[stream_elements]>=0.1.113,<0.2.0  # Added for StreamElementsSynthesizer

pydub>=0.25.1,<0.26.0

langchain-core>=0.3.10
langchain-classic

aiohttp==3.10.10

streamlit


filelock== 3.20.0


asyncpg == 0.30.0

aiosmtplib 


# libretranslate==1.3.9  # FREE self-hosted translation
# fasttext-langdetect==1.0.5  # Fast, accurate language detection (no external dependencies)

# langdetect==1.0.9

# Alternative: DeepL (optional paid backup)
deepl==1.23.0  # Uncomment if using DeepL as fallback





requests>=2.32.4
pandas==2.1.4
plotly==5.18.0
python-dateutil==2.8.2